baselines_iclr/requirements.txt
absl-py @ file:///croot/absl-py_1666362940888/work
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
atari-py==0.2.9
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
-e git+https://github.com/openai/baselines.git@ea25b9e8b234e6ee1bca43083f8f3cf974143998#egg=baselines
box2d-py==2.3.8
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
clikit==0.6.2
cloudpickle==1.2.1
crashtest==0.3.1
cycler==0.11.0
Cython==0.29.32
data==0.4
decorator==5.1.1
dm-reverb==0.8.0
dm-tree==0.1.7
docker-pycreds==0.4.0
docutils==0.18.1
dpcpp-cpp-rt==2022.2.0
exceptiongroup==1.1.1
Farama-Notifications==0.0.4
fasteners==0.17.3
flatbuffers==1.12
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.37.1
frozenlist==1.3.3
funcsigs==1.0.2
future==0.18.2
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.4
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.47.0
gym==0.15.7
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.7.0
httpstan==4.6.1
idna==3.3
ImageHash==4.3.1
imageio==2.21.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==2.0.0
intel-cmplr-lib-rt==2022.2.0
intel-cmplr-lic-rt==2022.2.0
intel-opencl-rt==2022.2.0
intel-openmp==2022.2.0
ipdb==0.13.13
ipython==7.34.0
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
joblib==0.14.1
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
latex==0.7.0
libclang==14.0.6
lockfile==0.12.2
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.1.1
matplotlib-inline==0.1.6
mkl==2022.2.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mpi4py==3.1.3
mpmath==1.2.1
-e git+https://github.com/openai/mujoco-py@a13903b82f1ab316815e63b3575526005b3b2ae1#egg=mujoco_py
multidict==6.0.4
numpy==1.21.6
numpydoc==1.5.0
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
parso==0.8.3
pastel==0.2.1
patchelf==0.17.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
portpicker==1.5.2
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
pyglet==1.5.0
Pygments==2.15.1
pylev==1.4.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pysimdjson==3.2.0
pystan==3.3.0
pytest==7.3.1
pytest-instafail==0.3.0
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rlds==0.1.5
rsa==4.9
-e git+https://github.com/geraudnt/skill-machines@131c8db75dddc9de5c60791427e0f0cbfa360d3d#egg=safety_gym&subdirectory=src/safety-gym
scipy==1.7.3
seaborn==0.8.1
sentry-sdk==1.21.1
setproctitle==1.3.2
shutilwhich==1.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
-e git+https://github.com/openai/spinningup.git@038665d62d569055401d91856abb287263096178#egg=spinup
stable-baselines3==2.0.0
sympy==1.10.1
tbb==2021.7.0
tempdir==0.7.1
tensorboard==2.0.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
traitlets==5.9.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3==1.26.11
wandb==0.15.1
wcwidth==0.2.6
webargs==8.1.0
Werkzeug==2.2.2
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.9.2
zipp @ file:///croot/zipp_1672387121353/work
